crud de compras

+++ database/migrations/2023_11_31_024921_create_shoppings_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->date('date')->comment('fecha de la compra');
             $table->string('code', 50)->comment('factura');
-            $table->unsignedInteger('total')->comment('total de la compra');
-            $table->enum('status', [1,0])->default(1);
+            $table->double('total', 8, 2)->comment('total de la compra');
+            $table->enum('status', [1, 0])->default(1);
             $table->foreignId('user_id')->constrained()->cascadeOnDelete()->cascadeOnUpdate()->comment('id del usuario');
             $table->unsignedInteger('provider_id')->comment('id del proveedor');
             $table->foreign('provider_id')->references('id')->on('providers')->cascadeOnDelete()->cascadeOnUpdate()->comment('id del proveedor');
